code rev

+++ kernel/src/kernel.c
@@ -115,8 +115,7 @@ __attribute__((section(".text._start_entry"))) noreturn void _start_entry()
         con_col_t old_col = CON_getConsoleColor();
         CON_setConsoleColor2(VGA_COLOR_RED, VGA_COLOR_BRIGHT_CYAN);
         volatile boot_MEM_MAP_Info_Entry_t* mem_map = MEM_MAP_ENTRY_PTR(_sys_info);
-        
-        for(int i = 0; i < (int)_sys_info->num_mem_map_entries; i++)
+        for(int i = 0; i < _sys_info->num_mem_map_entries; i++)
         {
             CON_printf("test");
             const char* mem_types[] = {
